+++ google/src/test/scala/org/broadinstitute/dsde/workbench/google/GoogleUtilitiesSpec.scala
@@ -107,6 +107,10 @@ class GoogleUtilitiesSpec
     when409(buildGoogleJsonResponseException(409)) shouldBe true
 
     whenGroupDoesNotExist(buildGoogleJsonResponseException(400, Some("does not exist"), None, None)) shouldBe true
+
+    when412(buildGoogleJsonResponseException(412, None, Some("conditionNotMet"), Some("global"))) shouldBe true
+    when412(buildGoogleJsonResponseException(409)) shouldBe false
+    when412(new NoSuchElementException()) shouldBe false
   }
 
   it should "return false in negative cases" in {