+++ google2/CHANGELOG.md
@@ -2,9 +2,38 @@
 
 This file documents changes to the `workbench-google2` library, including notes on how to upgrade to new versions.
 
+## 0.34
+
+SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.34-2147824"`
+
+
+### Changes
+* Added ability for getBucket to log at WARN level instead of ERROR
+* Updated GCE, Dataproc, and GKE interpreters to use `IO.blocking` combinator for blocking API calls through 
+the Google SDK.
+* Add `patchReplicas` to `KubernetesService`
+* Add `listDeployments` to `KubernetesService`
+
+### Dependency upgrades
+| Dependency                    |     Old Version      |          New Version |
+|-------------------------------|:--------------------:|---------------------:|
+| google-cloud-pubsub           |       1.125.2        |             1.125.11 |
+| google-cloud-nio              |       1.127.3        |              1.127.7 |
+| google-cloud-bigquery         |        2.32.1        |               2.34.1 |
+| google-api-services-container | v1-rev20230809-2.0.0 | v1-rev20231024-2.0.0 |
+| google-cloud-storage          |        2.27.0        |               2.29.1 |
+| google-cloud-dataproc         |        4.21.0        |               4.27.0 |
+| google-cloud-compute          |        1.34.0        |               1.40.0 |
+| google-cloud-container        |        2.27.0        |               2.33.0 |
+| google-cloud-billing          |        2.24.0        |               2.30.0 |
+| cats-mtl                      |        1.3.1         |                1.4.0 |
+| google-cloud-storage-transfer | 1.24.0 |               1.30.0 |
+| google-cloud-resourcemanager | 1.26.0 |               1.32.0 |
+
+
 ## 0.33
 
-SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.33-7362eef"`
+SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.33-e761452"`
 
 ### Changes
 Upgrade google-cloud-pubsub from 1.124.1 to 1.124.2
@@ -13,16 +42,23 @@ Upgrade google-cloud-pubsub from 1.124.1 to 1.124.2
 ### Dependency upgrades
 | Dependency   | Old Version | New Version |
 |----------|:-----------:|------------:|
-| google-cloud-pubsub |   1.124.1   |     1.124.2 |
-| google-cloud-nio |  0.126.19   |     0.127.2 |
+| google-cloud-pubsub |   1.124.1   |     1.125.2 |
+| google-cloud-nio |  0.126.19   |     1.127.3 |
 | grpc-core |   1.57.2    |      1.58.0 |
-| google-cloud-bigquery |   2.31.1    |      2.31.2 |
+| google-cloud-bigquery |   2.31.1    |      2.32.0 |
 | google-api-services-container |   v1-rev20230724-2.0.0    |      v1-rev20230809-2.0.0 |
 | sbt-scoverage |    2.0.8    |       2.0.9 |
 | scalatest |   3.2.16    |      3.2.17 |
 | scala       |   2.13.11   |     2.13.12 |
 | sbt-scalafix       |   0.11.0    |      0.11.1 |
-
+| google-cloud-storage         |  2.24.0     |      2.27.0 |
+| google-cloud-kms          |  2.26.0     |      2.27.0 |
+| google-cloud-dataproc          |  4.20.0    |      4.21.0 |
+| google-cloud-compute          |  1.33.0    |      1.34.0 |
+| google-cloud-container          |  2.26.0     |      2.27.0 |
+| google-cloud-billing          |  2.23.0     |      2.24.0 |
+| google-cloud-storage-transfer          |  1.23.0    |      1.24.0 |
+| google-cloud-resourcemanager          |  1.25.0    |      1.26.0 |
 
 ## 0.32